Ex 4.21

+++ bloglist/controllers/blogs.js
@@ -56,9 +56,23 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  if (await Blog.findByIdAndRemove(request.params.id))
-    response.status(204).end()
-  else response.status(400).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  console.log(user)
+  console.log(blog)
+
+  if (blog.user.toString() === user.id.toString()) {
+    if (await Blog.findByIdAndRemove(request.params.id))
+      response.status(204).end()
+    else response.status(400).end()
+  } else {
+    response.status(401).end()
+  }
 })
 
 module.exports = blogsRouter
